refactor: eval result by grid table with preview
