+++ console/packages/starwhale-core/src/dataset/index.ts
@@ -1,4 +1,3 @@
 export * from './types'
 export * from './utils'
-export * from './hooks/useDatasets'
 export * from './hooks/useDatasetTableAnnotations'
