+++ console/packages/starwhale-core/src/dataset/utils.ts
@@ -36,9 +36,17 @@ export const isAnnotationType = (type: string) => {
 export const isAnnotation = (data: any) => (typeof data === 'object' && isAnnotationType(data?._type)) || isMask(data)
 export const isAnnotationHiddenInTable = (data: any) => isAnnotation(data) && !isMask(data)
 
+// from: _owner "http://controller:8082/project/starwhale/dataset/pfp/version/yqpjypceemtxqhk45oy5wjuyfnf3z44gv5pez5tr"
+// to: e2e-20230516/controller/project/1/common-dataset/pfp/03/032e234c996e43a4dbfa1a7936864cc59bafa6465ce1ff6a4a10935c0defb84c375ecc6d32ea78a3ab4451fbed05057899dc66bc2cce1824c8eebc0ce240d7fa
+export function linkWithOwner(data: ITypeLink): string {
+    const matches = data._owner.match(/project\/(.*?)\/dataset\/(.*?)\/version\/(.*)/)
+    const [, projectName, datasetName, datasetVersionName] = matches ?? []
+    const token = (window.localStorage && window.localStorage.getItem('token')) ?? ''
+    return parseDataSrc(projectName, datasetName, datasetVersionName, token, data)
+}
 export function linkToData(data: ITypeLink, curryParseLinkFn: any): string {
+    if (data._owner) return linkWithOwner(data)
     if (data.uri.startsWith('http')) return data.uri
-    if (data._owner) return data._owner
     if (curryParseLinkFn) return curryParseLinkFn(data)
     return data.uri
 }
@@ -66,6 +74,7 @@ export function getSummary(record: RecordT, options: OptionsT) {
                         _extendSrc: anno.link ? linkToData(anno.link, options.parseLink) : undefined,
                         _extendType: isMask(anno) ? AnnotationType.MASK : anno._type,
                     })
+                    if (anno._owner) console.log(anno._owner, anno.link, [anno._owner, anno.link?.uri].join('/'))
                 }
 
                 Object.entries(anno).forEach(([_key, tmp]: any) => {
