+++ console/packages/starwhale-widgets/src/PanelTableWidget/index.tsx
@@ -38,6 +38,7 @@ function PanelTableWidget(props: WidgetRendererProps<any, any>) {
             records={data.records}
             storeKey={id}
             queryinline
+            previewable
             onCurrentViewChange={onCurrentViewChange}
             onInit={onInit}
         />
